+++ storybook/CHANGELOG.md
@@ -3,6 +3,101 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.1](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-components-stories@1.5.0...@equinor/fusion-react-components-stories@1.5.1) (2021-05-26)
+
+**Note:** Version bump only for package @equinor/fusion-react-components-stories
+
+
+
+
+
+# [1.5.0](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-components-stories@1.4.0...@equinor/fusion-react-components-stories@1.5.0) (2021-05-26)
+
+
+### Bug Fixes
+
+* fixed height ([5a5cf06](https://github.com/equinor/fusion-react-components/commit/5a5cf06e53b788789785a0ab33f5197356d77083))
+* fixed multiple bugs and added label ([676ebc7](https://github.com/equinor/fusion-react-components/commit/676ebc73d65d1e9582d2943e3e74ad74e7d8ae1c))
+
+
+### Features
+
+* added support for selection of date range ([751abbc](https://github.com/equinor/fusion-react-components/commit/751abbcdb9a4a7aa5478a82158b71b6ca188b1f6))
+
+
+
+
+
+# [1.4.0](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-components-stories@1.3.1...@equinor/fusion-react-components-stories@1.4.0) (2021-05-25)
+
+
+### Features
+
+* added support for selection of date range ([751abbc](https://github.com/equinor/fusion-react-components/commit/751abbcdb9a4a7aa5478a82158b71b6ca188b1f6))
+
+
+
+
+
+## [1.3.1](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-components-stories@1.3.0...@equinor/fusion-react-components-stories@1.3.1) (2021-05-18)
+
+**Note:** Version bump only for package @equinor/fusion-react-components-stories
+
+
+
+
+
+# [1.3.0](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-components-stories@1.2.1...@equinor/fusion-react-components-stories@1.3.0) (2021-05-11)
+
+
+### Bug Fixes
+
+* fixes to  header and icon handling ([babc3bd](https://github.com/equinor/fusion-react-components/commit/babc3bdd70714e02d205aad59e8ecc2d99824c36))
+
+
+### Features
+
+* added arrow icon to dropdown and other minor fixes ([03a8a84](https://github.com/equinor/fusion-react-components/commit/03a8a84ccf6d080bd65ae325748239d5c967197b))
+
+
+
+
+
+## [1.2.1](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-components-stories@1.2.0...@equinor/fusion-react-components-stories@1.2.1) (2021-05-04)
+
+
+### Bug Fixes
+
+* added new variations to datepicker and other small fixes ([d3bbdee](https://github.com/equinor/fusion-react-components/commit/d3bbdee0e80cb059b1340a1efa1e355f0877db5b))
+
+
+
+
+
+# [1.2.0](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-components-stories@1.1.3...@equinor/fusion-react-components-stories@1.2.0) (2021-05-04)
+
+
+### Bug Fixes
+
+* fixed reference to local yalc package ([ce1d3ee](https://github.com/equinor/fusion-react-components/commit/ce1d3eea22581616c00944d2ce64623ea3f52ebd))
+
+
+### Features
+
+* new datepicker component ([c06eee4](https://github.com/equinor/fusion-react-components/commit/c06eee4636e43c2f0c2752c1ce37436851dff487))
+
+
+
+
+
+## [1.1.3](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-components-stories@1.1.2...@equinor/fusion-react-components-stories@1.1.3) (2021-04-28)
+
+**Note:** Version bump only for package @equinor/fusion-react-components-stories
+
+
+
+
+
 ## [1.1.2](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-components-stories@1.1.1...@equinor/fusion-react-components-stories@1.1.2) (2021-04-13)
 
 **Note:** Version bump only for package @equinor/fusion-react-components-stories